% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{MarkerIntensityScan}
\alias{MarkerIntensityScan}
\title{MarkerIntensityScan}
\usage{
MarkerIntensityScan(gene, data, method)
}
\arguments{
\item{gene}{Marker gene of interest}

\item{data}{scRNAseq S4 Seurat object of single cell data after normalization}

\item{method}{Normalization methods and outlier detection of choice. 'SCTransform', 'RandomForest', 'scone'}
}
\value{
A vector of barcodes of cells with outlier marker gene expression
}
\description{
Scan all cells to detect those with high levels of a chosen marker gene
}
\examples{
outliers <- CellMarkerScan("IFNG",data,method="scone");
}
